# Serialization optimizations for .NET Core 3.1