String serialization:

```md
|   Method |      Mean |  Gen 0 | Allocated |
|--------- |----------:|-------:|----------:|
|   Before |  94.83 ns | 0.0194 |     112 B |
|    After |  68.11 ns |      - |         - |
```

String deserialization:

``` md
|   Method |      Mean |  Gen 0 | Allocated |
|--------- |----------:|-------:|----------:|
|   Before | 134.18 ns | 0.0527 |     304 B |
|    After | 109.08 ns | 0.0334 |     192 B |
```