@ReubenBond 2020-01-01
Ok, it was an edge case - the output buffer needs to have *some* space or there'll be an exception.